requirements.txt
Django==3.0.1
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==5.0.0
psycopg2==2.7.5
redis==2.10.6
celery==4.2.1
pyScss==1.3.5
django-autocomplete-light==3.3.2
django-ckeditor==5.6.1
django-constance==2.3.1
django-debug-toolbar==1.10.1
django-jinja==2.4.1
django-versatileimagefield==1.10
django-silk==3.0.1
django-redis==4.9.0
django-celery-email==2.0.1
django-anymail==4.2
django-post-office==3.1.0
